fix: precision loss for xprotocol clientStreamIDBase (#2404)
* fix: precision loss for xprotocol clientStreamIDBase

* fix: file header of license

+++ pkg/protocol/xprotocol/bolt/protocol.go
@@ -224,5 +224,6 @@ func (proto boltProtocol) EnableWorkerPool() bool {
 }
 
 func (proto boltProtocol) GenerateRequestID(streamID *uint64) uint64 {
-	return atomic.AddUint64(streamID, 1)
+	// fix bug for issue: https://github.com/mosn/mosn/issues/2403
+	return uint64(uint32(atomic.AddUint64(streamID, 1)))
 }
